pyproject.toml
[tool.poetry]
name = "pypl2mp3"
version = "0.1.0"
description = "YouTube playlist MP3 converter"
authors = ["Thierry Thiers <webcoder31@gmail.com>"]
license = "CeCILL-C"
readme = "README.md"
packages = [
  { include = "pypl2mp3", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
mutagen = "^1.47.0"
pygame = "^2.6.0"
colorama = "^0.4.6"
moviepy = "^1.0.3"
proglog = "^0.1.10"
shazamio = "^0.4.0"
python-slugify = "^8.0.4"
sshkeyboard = "^2.3.1"
thefuzz = "^0.22.1"
pytubefix = "^8.12.1"
rich-argparse = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pypl2mp3 = "pypl2mp3.main:main"
